pyproject.toml
[tool.poetry]
name = "agi-mcp-agent"
version = "0.1.0"
description = "An intelligent agent framework based on Master Control Program architecture"
authors = ["OT2.net <info@ot2.net>"]
readme = "README.md"
repository = "https://github.com/ot2net/agi-mcp-agent"
license = "MIT"
packages = [{include = "agi_mcp_agent"}]

[tool.poetry.dependencies]
python = ">=3.9,<3.12"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
pydantic = "^2.4.2"
sqlalchemy = "^2.0.22"
langchain = "^0.0.335"
numpy = "^1.24.3"
python-dotenv = "^1.0.0"
openai = "^1.2.4"
tiktoken = "^0.5.1"
requests = "^2.31.0"
aiohttp = "^3.11.18"
beautifulsoup4 = "^4.13.4"
websockets = "^11.0.3"
python-multipart = "^0.0.6"
anthropic = "^0.8.0"
mistralai = "^0.0.9"
google-generativeai = "^0.3.0"
alembic = "^1.11.1"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.3"
pytest-cov = "^4.1.0"
black = "^23.10.1"
isort = "^5.12.0"
flake8 = "^6.1.0"
mypy = "^1.6.1"
pre-commit = "^3.5.0"
sphinx = "^7.2.6"
sphinx-rtd-theme = "^1.3.0"
pytest-asyncio = "<0.26.0"

[tool.poetry.scripts]
agi-mcp-agent-api = "agi_mcp_agent.api.server:start_server"
llm-example = "examples.llm_example:main"
db-init = "scripts.db_init:init_database"
db-test = "scripts.db_test:test_database"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py39']

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false 